Add getTotalXXXRecipes functions to RecipeDAO, to be able to give a total page count for each recipe page view element.
Add these functions to interface and recipe test page as well.

Add new private RecipeDAO class function getSearchColumns to be
able to reuse the bit of code that creates the part of the SQL query
that applies the user search query to specific DB columns.

Add false return type to getSearchRecipes.